Fixed some bugs in the extension of the mesh. Solved the issue of extension and remeshing within one time step.

+++ src/controller.py
@@ -342,7 +342,7 @@ class Controller:
                         print("Remeshing by extending in the vertical direction the domain...")
 
                         # AM:
-                        elems_add = self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor * 2 - 1)
+                        elems_add = int(self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor * 2 - 1))
                         if elems_add % 2 != 0:
                             elems_add = elems_add + 1
 
@@ -375,7 +375,7 @@ class Controller:
                         print("Remeshing by extending in the horizontal direction the domain...")
 
                         # AM:
-                        elems_add = self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor * 2 - 1)
+                        elems_add = int(self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor * 2 - 1))
                         if elems_add % 2 != 0:
                             elems_add = elems_add + 1
 
@@ -410,7 +410,7 @@ class Controller:
                             print("Remeshing by extending towards negative y, positive y hit also so extending by half "
                                   "the factor in vertical direction...")
                             # AM:
-                            elems_add = self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -422,7 +422,7 @@ class Controller:
                             print("Remeshing by extending towards negative y...")
 
                             # AM:
-                            elems_add = self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -457,7 +457,7 @@ class Controller:
                             print("Remeshing by extending towards positive y, negative y hit also so extending by half "
                                   "the factor in vertical direction...")
                             # AM:
-                            elems_add = self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -469,7 +469,7 @@ class Controller:
                             print("Remeshing by extending towards positive y...")
 
                             # AM:
-                            elems_add = self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.ny * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -504,7 +504,7 @@ class Controller:
                             print("Remeshing by extending towards negative x, positive x hit also so extending by half "
                                   "the factor in vertical direction...")
                             # AM:
-                            elems_add = self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -516,7 +516,7 @@ class Controller:
                             print("Remeshing by extending towards negative x...")
 
                             # AM:
-                            elems_add = self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -551,7 +551,7 @@ class Controller:
                             print("Remeshing by extending towards positive x, negative x hit also so extending by half "
                                   "the factor in vertical direction...")
                             # AM:
-                            elems_add = self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
 
@@ -563,7 +563,7 @@ class Controller:
                             print("Remeshing by extending towards positive x...")
 
                             # AM:
-                            elems_add = self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1)
+                            elems_add = int(self.fracture.mesh.nx * (self.sim_prop.meshExtensionFactor - 1))
                             if elems_add % 2 != 0:
                                 elems_add = elems_add + 1
